requirements.txt
streamlit
python-dotenv
openai-agents
httpx
pandas
numpy
altair

